ui: home: move dir from /main
