+++ server/config/server.js
@@ -19,9 +19,7 @@ function koaConfig () {
     try {
       ctx.state.user = await firebaseConfig.verifyIdToken(ctx.headers.authorization.substring(7)/* strip 'Bearer ' prefix */)
     } catch (err) {
-      if (config.app.debugMode) {
-        console.debug(err)
-      }
+      config.app.debugMode && console.debug(err)
       ctx.throw(401, 'Invalid authorization token.')
     }
     return next()
@@ -35,12 +33,8 @@ function koaConfig () {
     try {
       await next()
     } catch (err) {
-      if (config.app.debugMode && (err instanceof AssertionError || err.expose)) {
-        console.debug(err)
-      }
-      if (err instanceof AssertionError) {
-        ctx.throw(400, err.message)
-      }
+      config.app.debugMode && (err instanceof AssertionError || err.expose) && console.debug(err)
+      err instanceof AssertionError && ctx.throw(400, err.message)
       throw err
     }
   })
