+++ src/App.js
@@ -10,8 +10,8 @@ import './comps/global/FontAwesome'
 import config from './data/config'
 import server from './data/server'
 import Navbar from './comps/global/Navbar'
-import Home from './comps/main/Home'
-import Login from './comps/main/Login'
+import Home from './comps/home/Home'
+import Login from './comps/home/Login'
 import Profile from './comps/account/Profile'
 import Shop from './comps/shop/Shop'
 import ShopItem from './comps/shop/ShopItem'
